requirements.txt
flask
flask-restful
python-dotenv
requests
firebase-admin